refactor: add resources folder
