+++ packages/wallet-connector-solana/generatePredicateResources.ts
@@ -36,5 +36,5 @@ function base64ToUint8Array(base64: string) {
 }
 `;
 
-fs.writeFileSync(`${__dirname}/src/predicateResources.ts`, code);
+fs.writeFileSync(`${__dirname}/src/resources/predicateResources.ts`, code);
 console.log('Generated'); // eslint-disable-line no-console
