+++ packages/wallet-connector-solana/src/resources/index.ts
@@ -1,0 +1,2 @@
+export * from "./predicateResources";
+export * from "./scriptResources";
