+++ packages/wallet-connector-solana/src/resources/predicateResources.ts
