+++ packages/wallet-connector-solana/src/resources/scriptResources.ts
